+++ src/components/filePage/FileHeader.tsx
@@ -3,7 +3,8 @@ import { FiSearch } from 'react-icons/fi';
 import fileIcon from '../images/Group 771.png';
 import { useRecoilState } from 'recoil';
 import { searchFileState, IsSearchState } from 'state';
-import { SearchFileInfo, SearchInputProps } from 'interfaces';
+import { SearchInputProps } from 'interfaces/fileType';
+import { SearchFileInfo } from 'interfaces/fileType';
 import back from '../images/Vector.png';
 import { useNavigate } from 'react-router-dom';
 import * as Style from '../../css/FilePage/FileHeaderStyle';
