+++ src/components/popup/TeamMateInfo.tsx
@@ -3,7 +3,8 @@ import styled from 'styled-components';
 import prof from '../images/template1.png';
 import { FiLink2 } from 'react-icons/fi';
 import axios from 'axios';
-import { ITeamMate, StyledProfileImgInfo } from '../../interfaces';
+import { ITeamMate } from 'interfaces/teamType';
+import { StyledProfileImgInfo } from 'interfaces/userType';
 import { useRecoilState } from 'recoil';
 import { teamMateNumState, modal2State, teamidState } from 'state';
 import { confirmAlert } from 'react-confirm-alert';
