+++ src/components/teampleDetailPage/detailBox.tsx
@@ -10,7 +10,8 @@ import deleteBtn from '../images/delete.png';
 import ellipse from '../images/Ellipse 1.png';
 import Send from '../images/send.png';
 import axios from 'axios';
-import { detailInfo, userInfo } from 'interfaces';
+import { detailInfo } from 'interfaces/taskType';
+import { userInfo } from 'interfaces/userType';
 import S3 from 'react-aws-s3-typescript';
 import { config } from 'config';
 import { useRecoilState } from 'recoil';
