+++ src/interfaces/modalType/index.ts
@@ -1,0 +1,4 @@
+export interface ModalProps {
+  setModal: React.Dispatch<React.SetStateAction<boolean>>;
+  setNextModal: React.Dispatch<React.SetStateAction<boolean>>;
+}
