+++ src/interfaces/stageType/index.ts
@@ -1,0 +1,6 @@
+export interface stageInfo {
+  name: string;
+  sequenceNum: number;
+  dueDate: Date | string;
+  startDate: Date | string;
+}
