+++ src/interfaces/userType/index.ts
@@ -1,0 +1,13 @@
+export interface StyledProfileImgInfo {
+  profileImage: string;
+}
+
+export interface userInfo {
+  email: string;
+  entranceYear: string;
+  major: string;
+  name: string;
+  profileImage: string;
+  schoolName: string;
+  subscribePlan: string;
+}
